google_interview.cpp
This program is my solution to a google job interview question I saw.  The question presented a small town in need of a mayor, the mayor must be a person that everyone knows.  Person A may know person B, but Person B may not know Person A.  My solution assumes that the town is small and that everyone can name who they know, it also assumes that everyone is equally social and knows the same number of other people.  It initializes a list of voters with their names, and the names of whomever they know.  It then checks each person a given voter knows and adds a vote to each person the voter knows.  If any person is known by everyone else they will be declared eligible to be mayor by the program.